add: remind fine push API & push API trigger to kafka event
