+++ moit-api/src/main/kotlin/com/mashup/moit/scheduler/NotiPushScheduler.kt
@@ -1,11 +1,15 @@
 package com.mashup.moit.scheduler
 
+import com.mashup.moit.domain.fine.FineService
 import com.mashup.moit.domain.moit.MoitService
 import com.mashup.moit.domain.study.StudyService
+import com.mashup.moit.domain.user.UserService
 import com.mashup.moit.infra.event.EventProducer
+import com.mashup.moit.infra.event.RemindFineNotificationPushEvent
 import com.mashup.moit.infra.event.ScheduledStudyNotificationPushEvent
 import com.mashup.moit.infra.event.StudyAttendanceStartNotificationPushEvent
 import com.mashup.moit.infra.fcm.FCMNotificationService
+import com.mashup.moit.infra.fcm.FineRemindNotification
 import com.mashup.moit.infra.fcm.ScheduledStudyNotification
 import com.mashup.moit.infra.fcm.StudyAttendanceStartNotification
 import org.slf4j.Logger
@@ -18,7 +22,9 @@ import java.time.LocalDateTime
 @Component
 class NotiPushScheduler(
     private val studyService: StudyService,
+    private val userService: UserService,
     private val moitService: MoitService,
+    private val fineService: FineService,
     private val fcmNotificationService: FCMNotificationService,
     private val eventProducer: EventProducer
 ) {
@@ -46,7 +52,7 @@ class NotiPushScheduler(
 
         logger.info("Done Push notification for {} studies, at {}.", startedStudy.size, LocalDateTime.now())
     }
-    
+
     @Scheduled(cron = "0 */5 * * * *")
     @Async("pushSchedulerExecutor")
     fun pushScheduledStudyRemindNotification() {
@@ -67,5 +73,48 @@ class NotiPushScheduler(
 
         studies.forEach { study -> studyService.markAsPushed(study.id) }
         logger.info("Done Push notification for {} scheduled studies, at {}.", studyMoitMap.size, LocalDateTime.now())
+    }
+
+    @Scheduled(cron = "0 30 10 * * *")
+    @Async("pushSchedulerExecutor")
+    fun pushRemindFineNotification() {
+        val scheduleContext = org.joda.time.LocalDateTime.now()
+
+        val unrequestedFines = fineService.getUnrequestedFines()
+        logger.info("Remind {} fines! Start Push Notification at {}.", unrequestedFines.size, scheduleContext)
+
+        val unrequestedFineUserMap = userService
+            .findUsersById(
+                unrequestedFines.map { fine -> fine.userId }.distinct()
+            )
+            .associateBy { user -> user.id }
+        val unrequestedFineMoitMap = moitService
+            .getMoitsByIds(
+                unrequestedFines.map { fine -> fine.moitId }.toSet()
+            )
+            .associateBy { moit -> moit.id }
+        val unrequestedFineStudyMap = studyService
+            .findByStudyIds(
+                unrequestedFines.map { fine -> fine.studyId }.distinct()
+            )
+            .associateBy { study -> study.id }
+
+        for (fine in unrequestedFines) {
+            val user = unrequestedFineUserMap[fine.userId] ?: continue
+            val moit = unrequestedFineMoitMap[fine.moitId] ?: continue
+            val study = unrequestedFineStudyMap[fine.studyId] ?: continue
+            FineRemindNotification.of(
+                user = user,
+                moit = moit,
+                study = study
+            )?.let { notification -> fcmNotificationService.pushRemindFineNotification(notification) }
+        }
+        eventProducer.produce(
+            RemindFineNotificationPushEvent(
+                fineIds = unrequestedFines.map { fine -> fine.id }.toSet(),
+                flushAt = LocalDateTime.now()
+            )
+        )
+        logger.info("Done Push notification for {} fines, at {}", unrequestedFines.size, org.joda.time.LocalDateTime.now())
     }
 }
